OneDrive/Desktop/new_excel_chatbot/requirements.txt
streamlit
pandas
matplotlib
seaborn
openpyxl
python-dotenv
requests
